Rewrite Qt6 compatibility code
Revert "Add a shim for (Qt|QString)::SkipEmptyParts"
This reverts commit 438cd4e5fee2de63d6a23a0aa2299f116337fee9.

Revert "Add rudimentary Qt6 support"
This reverts commit ed8e85ce14de901a5ee2373cc3532e60f7fea07d.

+++ src/desktop/scene/lasertrailitem.cpp
@@ -72,7 +72,7 @@ void LaserTrailItem::paint(QPainter *painter, const QStyleOptionGraphicsItem *op
 	painter->restore();
 }
 
-bool LaserTrailItem::animationStep(float dt)
+bool LaserTrailItem::animationStep(qreal dt)
 {
 	m_blink = std::fmod(m_blink + dt, ANIM_TIME);
